Created a game service and interface
